requirements.txt
flask==3.0.0
pillow==10.1.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
boto3==1.34.7
supabase==2.3.1
requests==2.31.0
